% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscfish.R
\docType{package}
\name{bscfish}
\alias{bscfish}
\alias{bscfish-package}
\title{bscfish: A package for basic fish life history functions}
\description{
The bscfish package contains an r file with three basic and typical life history
functions for fishes.
}
\section{bscfish functions}{

L_age
M_age
W_age
}

