% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_functions.R
\name{L_age}
\alias{L_age}
\title{TITLE: Von Bertalanffy Growth Curve.}
\usage{
L_age(L_inf, k, age, t_0)
}
\arguments{
\item{L_inf}{theoretical maximum average length}

\item{k}{growth coefficient}

\item{age}{age of fish. independent variable}

\item{t_0}{theoretical age at zero length}
}
\value{
Returns length of fish at specified age
}
\description{
DESCRIPTION: Determine fish length from age imput.
}
\details{
DETAILS: This section provides furthur information about the function.
}
\section{Section_Test}{

This section was written to test the commit and push features of git/github.
NOTE: The section title must be followed by a colon. This is similar to R-Markdown sections
}

\examples{
## Inputs from: https://www.niwa.co.nz/node/111765
pat_tooth_L_inf = 169.07
pat_tooth_k     = 0.093
pat_tooth_t_0   = -0.256
ages = seq(1, 30, 0.1)
multiple_L = L_age(169.07, 0.093, ages, -0.256)
plot(ages, multiple_L)
}
