% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_functions.R
\name{M_age}
\alias{M_age}
\title{Age at Maturity function}
\usage{
M_age(beta, age, age_50)
}
\arguments{
\item{beta}{Slope of the curve at \code{age == age_50}. Correct value in example?}

\item{age}{Age of the fish prob(maturity) is desired.}

\item{age_50}{Age at which 50 percent of fish are mature.}
}
\value{
Returns probability of maturity at specified age
}
\description{
Determine the logistic curve for the relationship between age of fish and probability that it is sexually
}
\details{
This function uses a standard logistic formulation for the estimation of the probability a Patagonia Toothfish
is sexually mature based on its age. It relies on species specific parameters \code{beta} and \code{age_50}.
}
\examples{
## Inputs from: https://www.niwa.co.nz/node/111765
pat_tooth_beta = .5
ages = seq(1, 30, 0.1)
pat_tooth_age_50 = 12.8
multiple_M = M_age(pat_tooth_beta, ages, pat_tooth_age_50)
plot(ages, multiple_M)
}
